enhance entity and model names
