+++ src/main/java/menuManegment/demo/menu/mapper/MegaMenuMapper.java
@@ -1,14 +1,14 @@
 package menuManegment.demo.menu.mapper;
 
-import menuManegment.demo.menu.entity.MenuManagement;
-import menuManegment.demo.menu.model.MenuManagementModel;
+import menuManegment.demo.menu.entity.MegaMenu;
+import menuManegment.demo.menu.model.MegaMenuModel;
 import org.mapstruct.Mapper;
 import org.mapstruct.NullValuePropertyMappingStrategy;
 import org.springframework.stereotype.Component;
 
 @Component
 @Mapper(nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.IGNORE)
-public interface MenuManagementMapper extends GenericMapper<MenuManagement, MenuManagementModel> {
+public interface MegaMenuMapper extends GenericMapper<MegaMenu, MegaMenuModel> {
 //@MapperConfig(
 //        nullValueCheckStrategy = NullValueCheckStrategy.ALWAYS
 //)
