+++ src/main/java/menuManegment/demo/menu/repository/MegaMenuRepository.java
@@ -1,0 +1,9 @@
+package menuManegment.demo.menu.repository;
+
+import menuManegment.demo.menu.entity.MegaMenu;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MegaMenuRepository extends GenericRepository<MegaMenu> {
+
+}
