--- src/main/java/menuManegment/demo/menu/repository/MenuManagementRepository.java
@@ -1,9 +1,0 @@
-package menuManegment.demo.menu.repository;
-
-import menuManegment.demo.menu.entity.MenuManagement;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MenuManagementRepository extends GenericRepository<MenuManagement> {
-
-}
