+++ src/main/java/menuManegment/demo/menu/service/MegaMenuService.java
@@ -1,0 +1,8 @@
+package menuManegment.demo.menu.service;
+
+import menuManegment.demo.menu.entity.MegaMenu;
+import menuManegment.demo.menu.model.MegaMenuModel;
+
+public interface MegaMenuService extends CRUD<MegaMenu, MegaMenuModel> {
+
+}
