--- src/main/java/menuManegment/demo/menu/service/MenuManagementService.java
@@ -1,8 +1,0 @@
-package menuManegment.demo.menu.service;
-
-import menuManegment.demo.menu.entity.MenuManagement;
-import menuManegment.demo.menu.model.MenuManagementModel;
-
-public interface MenuManagementService extends CRUD<MenuManagement, MenuManagementModel> {
-
-}
